python-server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
openai>=1.10.0
chromadb>=0.4.22
watchdog>=3.0.0
python-frontmatter>=1.0.1
httpx>=0.26.0
